svn merge -r 25621:25749 http://svn.civicrm.org/civicrm/branches/v3.1 including CRM-3507, CRM-3897, CRM-3902, CRM-4312, CRM-4321, CRM-4523, CRM-4575, CRM-4849, CRM-4924, CRM-4966, CRM-4971, CRM-4975, CRM-5101, CRM-5110, CRM-5259, CRM-5270, CRM-5305, CRM-5403, CRM-5423, CRM-5478, CRM-5514, CRM-5546, CRM-5552, CRM-5553, CRM-5555, CRM-5558, CRM-5564, CRM-5565, CRM-5566, CRM-5574, CRM-5578, CRM-5579, CRM-5580, CRM-5581, CRM-5582, CRM-5583, CRM-5586, CRM-5587, CRM-5588, CRM-5589, CRM-5590, CRM-5592, CRM-5593, CRM-5594, CRM-5595, CRM-5597, CRM-5600, CRM-5601, CRM-5602, CRM-5606 works/fixes

+++ CRM/Contact/Form/Merge.php
@@ -538,8 +538,10 @@ class CRM_Contact_Form_Merge extends CRM_Core_Form
             $dao->free();
 
             // delete the main contact's file
-            CRM_Core_BAO_File::delete($fileIds[$this->_cid], $this->_cid, $customId);
-
+            if ( !empty($fileIds[$this->_cid]) ) {
+                CRM_Core_BAO_File::delete($fileIds[$this->_cid], $this->_cid, $customId);
+            }
+            
             // move the other contact's file to main contact
             $sql = "UPDATE {$tableName} SET {$columnName} = {$fileIds[$this->_oid]} WHERE entity_id = {$this->_cid}";
             CRM_Core_DAO::executeQuery($sql, CRM_Core_DAO::$_nullArray);
